+++ tabbycat/results/views.py
@@ -371,9 +371,14 @@ class BasePublicNewBallotSetView(PublicTournamentPageMixin, BaseBallotSetView):
     def populate_objects(self):
         self.object = self.get_object() # must be populated before self.error_page() called
 
-        round = self.get_tournament().current_round
-        if round.draw_status != Round.STATUS_RELEASED or not round.motions_released:
-            return self.error_page(_("The draw and/or motions for the round haven't been released yet."))
+        tournament = self.get_tournament()
+        round = tournament.current_round
+        if round.draw_status != Round.STATUS_RELEASED:
+            return self.error_page(_("The draw for this round hasn't been released yet."))
+
+        if (tournament.pref('enable_motions') or tournament.pref('motion_vetoes_enabled')) \
+                and not round.motions_released:
+            return self.error_page(_("The motions for this round haven't been released yet."))
 
         try:
             self.debateadj = DebateAdjudicator.objects.get(adjudicator=self.object, debate__round=round)
@@ -390,6 +395,11 @@ class BasePublicNewBallotSetView(PublicTournamentPageMixin, BaseBallotSetView):
         if not self.debate.adjudicators.has_chair:
             return self.error_page(_("Your debate doesn't have a chair, so you can't enter results for it. "
                     "Please contact a tab room official."))
+
+        if not (tournament.pref('draw_side_allocations') == 'manual-ballot' and
+                tournament.pref('teams_in_debate') == 'two') and not self.debate.sides_confirmed:
+            return self.error_page(_("It looks like the sides for this debate haven't yet been confirmed, "
+                    "so you can't enter results for it. Please contact a tab room official."))
 
     def error_page(self, message):
         # This bypasses the normal TemplateResponseMixin and ContextMixin