+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 Change Log
 ==========
 
+2.0.1
+-----
+*Release date: TBD*
+
+- Fixed issue where results submission would crash if sides are unconfirmed
+- Fixed issue where scoresheets would not display properly for adjudicators who lack institutions
+- Fixed issue where the round history indicators in the Edit Adjudicators page would sometimes omit the "rounds ago" indicator
+
+
 2.0.0 (Iberian Lynx)
 --------------------
 *Release date: 13 November 2017*